Replace malloc_not_svelte with malloc_low_memory.
The malloc_not_svelte variable name is confusing and makes the
low memory config the default. Change this so that the default is
the regular allocator, and that malloc_low_memory is used to enable
the low memory allocator.

Update blueprint rules so that the default config is not the
low memory one.

Test: Verified scudo config is used by default.
Test: Verfified Android GO config uses the jemalloc low memory config.
Change-Id: I5b34c6f65e50fc3556a31d5a5b72509f4ff9ae75